Moving decode to internal/common (#204)
* Moving decode to internal/common

* 07Feb20-Package Changed to common
